backend/requirements.txt
easyocr==1.7.2
Flask==3.1.1
language_tool_python==2.9.4
numpy==2.2.0
opencv_python==4.12.0.88
Pillow==11.3.0
psutil==7.0.0
torch==2.7.1
ultralytics==8.3.168
Werkzeug==3.1.3
piper-tts
flask-cors